requirements.txt
# LLM & LangChain
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.10
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-milvus>=0.0.5
langchain-ollama>=0.0.5

# Vector Stores
pymilvus>=2.3.5

# LLM Providers
openai>=1.12.0
tiktoken>=0.6.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Web Scraping & Parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Bio Informatics
biopython>=1.83

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6 